+++ src/liteloader/api/ScheduleAPI.cpp
@@ -20,36 +20,36 @@ atomic_uint nextTaskId = 0;
 class ScheduleTaskData {
 public:
     enum class TaskType { Delay, Repeat, InfiniteRepeat };
-    unsigned int              taskId;
+    uint32_t              taskId;
     TaskType                  type;
-    long long                 leftTime, interval;
-    int                       count;
+    int64_t                 leftTime, interval;
+    int32_t                       count;
     std::function<void(void)> task;
     HMODULE                   handle;
 
     ScheduleTaskData(
         TaskType                  type,
         std::function<void(void)> task,
-        unsigned long long        delay,
-        unsigned long long        interval,
-        int                       count,
+        uint64_t        delay,
+        uint64_t        interval,
+        int32_t                       count,
         HMODULE                   handle
     )
-    : type(type), task(std::move(task)), leftTime((long long)delay), interval((long long)interval), count(count),
+    : type(type), task(std::move(task)), leftTime((int64_t)delay), interval((int64_t)interval), count(count),
       taskId(++nextTaskId), handle(handle){};
 
-    [[nodiscard]] inline unsigned int getTaskId() const { return taskId; }
+    [[nodiscard]] inline uint32_t getTaskId() const { return taskId; }
 
     inline bool operator>(const ScheduleTaskData& t) const { return leftTime > t.leftTime; }
 };
 
 std::vector<ScheduleTaskData> pendingTaskList{};
-std::vector<unsigned int>     pendingCancelList{};
+std::vector<uint32_t>     pendingCancelList{};
 bool                          pendingClear = false;
 
 class ScheduleTaskQueueType : public priority_queue<ScheduleTaskData, vector<ScheduleTaskData>, greater<>> {
 public:
-    bool remove(unsigned int taskId) {
+    bool remove(uint32_t taskId) {
         bool removed = false;
 
         for (size_t i = 0; i < c.size(); ++i)
@@ -160,7 +160,7 @@ public:
         }
     }
 
-    bool has(unsigned int taskId) {
+    bool has(uint32_t taskId) {
 
         return std::ranges::any_of(c, [taskId](ScheduleTaskData const& task) { return task.taskId == taskId; });
     }
@@ -170,9 +170,9 @@ ScheduleTaskQueueType taskQueue;
 
 
 namespace Schedule {
-ScheduleTask delay(std::function<void(void)> task, unsigned long long tickDelay, HMODULE handle) {
+ScheduleTask delay(std::function<void(void)> task, uint64_t tickDelay, HMODULE handle) {
     if (ll::globalRuntimeConfig.serverStatus >= ll::LLServerStatus::Stopping)
-        return {(unsigned)-1};
+        return {(uint32_t)-1};
     ScheduleTaskData sche(ScheduleTaskData::TaskType::Delay, std::move(task), tickDelay, -1, -1, handle);
     locker.lock();
     pendingTaskList.push_back(sche);
@@ -180,9 +180,9 @@ ScheduleTask delay(std::function<void(void)> task, unsigned long long tickDelay,
     return {sche.getTaskId()};
 }
 
-ScheduleTask repeat(std::function<void(void)> task, unsigned long long tickRepeat, int maxCount, HMODULE handle) {
+ScheduleTask repeat(std::function<void(void)> task, uint64_t tickRepeat, int32_t maxCount, HMODULE handle) {
     if (ll::globalRuntimeConfig.serverStatus >= ll::LLServerStatus::Stopping)
-        return {(unsigned)-1};
+        return {(uint32_t)-1};
     ScheduleTaskData::TaskType type =
         maxCount < 0 ? ScheduleTaskData::TaskType::InfiniteRepeat : ScheduleTaskData::TaskType::Repeat;
     ScheduleTaskData sche(
@@ -196,13 +196,13 @@ ScheduleTask repeat(std::function<void(void)> task, unsigned long long tickRepea
 
 ScheduleTask delayRepeat(
     std::function<void(void)> task,
-    unsigned long long        tickDelay,
-    unsigned long long        tickRepeat,
-    int                       maxCount,
+    uint64_t        tickDelay,
+    uint64_t        tickRepeat,
+    int32_t                       maxCount,
     HMODULE                   handle
 ) {
     if (ll::globalRuntimeConfig.serverStatus >= ll::LLServerStatus::Stopping)
-        return {(unsigned)-1};
+        return {(uint32_t)-1};
     ScheduleTaskData::TaskType type =
         maxCount < 0 ? ScheduleTaskData::TaskType::InfiniteRepeat : ScheduleTaskData::TaskType::Repeat;
     ScheduleTaskData sche(type, std::move(task), tickDelay, std::max(tickRepeat, 1ull), maxCount, handle);
@@ -214,7 +214,7 @@ ScheduleTask delayRepeat(
 
 ScheduleTask nextTick(std::function<void(void)> task, HMODULE handle) {
     if (ll::globalRuntimeConfig.serverStatus >= ll::LLServerStatus::Stopping)
-        return {(unsigned)-1};
+        return {(uint32_t)-1};
     ScheduleTaskData sche(ScheduleTaskData::TaskType::Delay, std::move(task), 1, -1, -1, handle);
     locker.lock();
     pendingTaskList.push_back(sche);
@@ -235,7 +235,7 @@ void EndScheduleSystem() {
 }
 
 
-ScheduleTask::ScheduleTask(unsigned int taskId) : taskId(taskId) {}
+ScheduleTask::ScheduleTask(uint32_t taskId) : taskId(taskId) {}
 
 bool ScheduleTask::cancel() const {
     locker.lock();
