+++ src/mc/network/packet/ClientCacheBlobStatusPacket.h
@@ -10,8 +10,8 @@
 class ClientCacheBlobStatusPacket : public ::Packet {
 
 public:
-    std::vector<unsigned long long> mMissingIds; // this+0x30
-    std::vector<unsigned long long> mFoundIds;   // this+0x48
+    std::vector<uint64_t> mMissingIds; // this+0x30
+    std::vector<uint64_t> mFoundIds;   // this+0x48
 
     // prevent constructor by default
     ClientCacheBlobStatusPacket& operator=(ClientCacheBlobStatusPacket const&) = delete;
