+++ src/mc/network/packet/ClientCacheMissResponsePacket.h
@@ -12,7 +12,7 @@ class ClientCacheMissResponsePacket : public ::Packet {
 
 public:
     std::vector<std::shared_ptr<ClientBlobCache::Server::Blob>> mWriteMissingContent;    // this+0x30
-    std::unordered_map<unsigned long long, std::string>         mReceivedMissingContent; // this+0x48
+    std::unordered_map<uint64_t, std::string>         mReceivedMissingContent; // this+0x48
 
     // prevent constructor by default
     ClientCacheMissResponsePacket& operator=(ClientCacheMissResponsePacket const&) = delete;
